% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_model_MLL}
\alias{BC_model_MLL}
\title{classification BC model}
\format{
list with different prediction models
}
\usage{
data(models_L_BC)
}
\description{
model used for BC predictions including cross-validation
}
\keyword{datasets}
