% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Counts}
\alias{Counts}
\title{Expression example data}
\format{
data.frame with gene expression data
}
\usage{
data(Counts)
}
\description{
Gene expression example data used for classification
}
\keyword{datasets}
