% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ID_conv}
\alias{ID_conv}
\title{Gene ID conversion file}
\format{
data.frame with symbol,ensemble, entrez and default expression values
}
\usage{
data(ID_conversion)
}
\description{
Database to match different gene IDs
}
\keyword{datasets}
