% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models_20}
\alias{models_20}
\title{classification model}
\format{
list with different prediction models
}
\usage{
data(models_L)
}
\description{
model used for predictions including cross-validation
}
\keyword{datasets}
