% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models_L_Immuno}
\alias{models_L_Immuno}
\title{classification immuno model}
\format{
list with different prediction models
}
\usage{
data(models_L_immuno)
}
\description{
model used for immuno predictions including cross-validation
}
\keyword{datasets}
