% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLCatcheR.R
\name{outFun}
\alias{outFun}
\title{Classifiction}
\usage{
outFun(Counts.file = NA, ID_class = "symbol", sep = "\\t")
}
\arguments{
\item{Counts.file}{count data}

\item{ID_class}{gene ids}

\item{sep}{file seperator}
}
\value{
data.frame containing class predictions
}
\description{
Prediction of B-ALL leukemia subtypes based on expression data
}
\examples{
outFun()

}
