% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svm0.3}
\alias{svm0.3}
\title{ALL subtype classifier}
\format{
Classifier object
}
\usage{
data(svm0.3_models)
}
\description{
Trained classifier to detect molecular subclass from gene expression data
}
\keyword{datasets}
