% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test expression data}
\format{
data.frame with human gene expression counts
}
\usage{
data(test_data)
}
\description{
Raw gene read counts of 195 patients obtained from transcriptomics
}
\keyword{datasets}
