% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrun.R
\name{outFun}
\alias{outFun}
\title{Classifiction}
\usage{
outFun(Counts.file, ID_class = "ensemble_ID", sep = "\\t")
}
\arguments{
\item{Counts.file}{count data}

\item{ID_class}{gene ids}

\item{sep}{file seperator}
}
\value{
data.frame containing class predictions
outFun(Counts = "StJude_data_missing_genes.tsv", ID_class = "ensemble_ID", sep = "\t")
}
\description{
Prediction of B-ALL leukemia subtypes based on expression data
}
